fizzbuzz: Simplify find-fizzbuzz-index-for
Instead of taking one param and two optional ones, take one mandatory,
and an optional seq of key-value pairs. The previous max and limit
parameters can now be given as named params instead, which makes it
easier if one only wants to specify the limit, with using the default
max value.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>